// src/types/beats/stellar.ts
export interface StellarLedgerTransactionsPreconditions {
  timebounds?: StellarLedgerTransactionsPreconditionsTimebounds;
  ledgerbounds?: StellarLedgerTransactionsPreconditionsLedgerbounds;
  min_account_sequence?: string;
}
// ...
export interface StellarLedgerTransactionsOperations {
  id: string;
  paging_token: string;
  transaction_successful: boolean;
  source_account: string;
  type: string;
  type_i: number;
  created_at: string;
  transaction_hash: string;
  amount: string;
  price: string;
  price_r: StellarLedgerTransactionsOperationsPriceR;
  buying_asset_type: string;
  selling_asset_type: string;
  selling_asset_code: string;
  selling_asset_issuer: string;
  offer_id: string;
  buying_asset_code: string;
  buying_asset_issuer: string;
  asset_type: string;
  asset_code: string;
  asset_issuer: string;
  from: string;
  to: string;
  path: StellarLedgerTransactionsOperationsPath[];
  source_amount: string;
  destination_min: string;
  source_asset_type: string;
  source_asset_code: string;
  source_asset_issuer: string;
  source_max: string;
  asset: string;
  claimants: StellarLedgerTransactionsOperationsClaimants[];
  balance_id: string;
  claimant: string;
  set_flags: number[];
  set_flags_s: string[];
  trustee: string;
  trustor: string;
  authorize: boolean;
  authorize_to_maintain_liabilities: boolean;
  clear_flags: number[];
  clear_flags_s: string[];
  account: string;
  into?: string;
  liquidity_pool_id?: string;
  reserves_max?: StellarLedgerTransactionsOperationsReservesMax[];
  min_price?: string;
  min_price_r?: StellarLedgerTransactionsOperationsMinPriceR;
  max_price?: string;
  max_price_r?: StellarLedgerTransactionsOperationsMaxPriceR;
  reserves_deposited?: StellarLedgerTransactionsOperationsReservesDeposited[];
  shares_received?: string;
  limit: string;
  function: string;
  parameters: StellarLedgerTransactionsOperationsParameters[];
  address: string;
  salt: string;
  asset_balance_changes: null | unknown[];
  starting_balance: string;
  funder: string;
  sponsored_id: string;
  signer_key: string;
  signer_weight: number;
  sponsor: string;
  begin_sponsor: string;
  name: string;
  value: string;
}

export interface StellarLedgerTransactionsFeeBumpTransaction {
  hash: string;
  signatures: string[];
}

export interface StellarLedgerTransactionsInnerTransaction {
  hash: string;
  signatures: string[];
  max_fee: string;
}

export interface StellarLedgerTransactions {
  id: string;
  paging_token: string;
  successful: boolean;
  hash: string;
  ledger: number;
  created_at: string;
  source_account: string;
  source_account_sequence: string;
  fee_account: string;
  fee_charged: string;
  max_fee: string;
  operation_count: number;
  envelope_xdr: string;
  result_xdr: string;
  result_meta_xdr: string;
  fee_meta_xdr: string;
  memo_type: string;
  signatures: string[];
  preconditions: StellarLedgerTransactionsPreconditions;
  operations: StellarLedgerTransactionsOperations[];
  memo: string;
  memo_bytes: string;
  fee_bump_transaction: StellarLedgerTransactionsFeeBumpTransaction;
  inner_transaction: StellarLedgerTransactionsInnerTransaction;
}

export interface StellarLedger {
  id: string;
  paging_token: string;
  hash: string;
  prev_hash: string;
  sequence: number;
  successful_transaction_count: number;
  failed_transaction_count: number;
  operation_count: number;
  tx_set_operation_count: number;
  closed_at: string;
  total_coins: string;
  fee_pool: string;
  base_fee_in_stroops: number;
  base_reserve_in_stroops: number;
  max_tx_set_size: number;
  protocol_version: number;
  header_xdr: string;
  transactions: StellarLedgerTransactions[];
  _network: string;
  _vm: string;
}